adds the hostname of device in pushed events
Adds the `source_name` field in the `attributes` of the pushed events.
The `source_name` contains the hostname (from uci) of the device
running this code.

This way it is possible to distinguish access points and/or process the
'connected' state more intelligently - to prevent spurious transitions
to 'away' once the station roams to another access point.